chore: split into components and hooks ClickMappings
