+++ frontend/src/metabase/dashboard/components/ClickMappings/ClickMappings.module.css
