+++ frontend/src/metabase/dashboard/components/ClickMappings/index.ts
@@ -1,0 +1,3 @@
+export * from "./ClickMappings";
+export * from "./utils";
+export * from "./hooks";
